This has a fix for the performance issue we were seeing, taking like 5 seconds to finish loading the page. Turned out to be using ob_start('ob_gzhandler') was causing the issue. I took it out and added custom compressing code that uses php's gzcompress. Looks good.
Also added site_name to the Profile object and view, added an owner property on the base resource class (AppResource), fixed a bug in FollowerResource that required a follower request to be authorized, implemented Keith DiSarno's suggestion for the follow request confirmation view, issue 1 in github, updated some of the views with additional markup, updated the profile index view markup, and created a filters plugin. Oh, added an RSS representation of the posts for RSS readers.